+++ Jade-Mainframe/Data/Product.swift
@@ -10,6 +10,7 @@ import Foundation
 struct Product: Identifiable {
     var id = UUID()
     let productName: String
-    let productPrice: Int
+    let productPrice: Double
     let imageName: String
+    
 }
